# backend/DevPilot_OpenAPI_Postman.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevPilot API",
    "description": "AI-powered SDLC automation platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents/project/{project_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Read Project Documents",
        "operationId": "read_project_documents_api_v1_documents_project__project_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response Read Project Documents Api V1 Documents Project  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Document",
        "operationId": "create_document_api_v1_documents_project__project_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Read Document",
        "operationId": "read_document_api_v1_documents__document_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/task/{task_id}/status": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Task Status",
        "description": "Poll the status of a Celery task.\nPossible states: PENDING | STARTED | SUCCESS | FAILURE",
        "operationId": "get_task_status_api_v1_documents_task__task_id__status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/project/{project_id}/generate-brd-from-file": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Generate BRD From File",
        "description": "Upload a .txt or .docx file, extract its text, create a document record, and dispatch a Celery task to generate a BRD from the extracted text.\n\nPoll the returned task_id via GET /documents/task/{task_id}/status.",
        "operationId": "generate_brd_from_file_api_v1_documents_project__project_id__generate_brd_from_file_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "title": "Title",
                    "description": "Optional document title (defaults to filename)"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "title": "File",
                    "description": "Upload a .txt or .docx file"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BRD generation task dispatched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "task_id": { "type": "string" },
                    "document_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – unsupported file type or empty content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/project/{project_id}/generate-test-cases-from-file": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Generate Test Cases From File",
        "description": "Upload a BRD document (.txt or .docx), extract its text, create a document record, and dispatch a Celery task to generate test cases from the BRD text.\n\nPoll the returned task_id via GET /documents/task/{task_id}/status.",
        "operationId": "generate_test_cases_from_file_api_v1_documents_project__project_id__generate_test_cases_from_file_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "title": "Title",
                    "description": "Optional document title (defaults to filename)"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "title": "File",
                    "description": "Upload a .txt or .docx BRD file"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test case generation task dispatched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "task_id": { "type": "string" },
                    "document_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – unsupported file type or empty content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocType",
            "default": "brd"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DocumentCreate"
      },
      "DocumentResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocType",
            "default": "brd"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "org_id",
          "project_id",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}